+++ src/chinese/segmentor/implementations/viterbi/segmentor_impl.h
@@ -37,7 +37,7 @@ class CWeight {
    typedef CScoreMap< CTwoWords, SCORE_TYPE > CTwoWordsMap;
    typedef CScoreMap< pair<CWord, int>, SCORE_TYPE > CWordIntMap;
    typedef CScoreMap< pair<CTwoWords, int>, SCORE_TYPE > CTwoWordsIntMap;
-   typedef CScoreMap< unsigned int, SCORE_TYPE > CIntMap;
+   typedef CScoreMap< unsigned long int, SCORE_TYPE > CIntMap;
 
 public:
    CWordMap m_mapSeenWords;
