+++ src/chinese/tagger/implementations/agendachart/tagger.h
@@ -29,7 +29,7 @@ protected:
    int m_nScoreIndex;
 
 public:
-   CTagger(string sFeatureDBPath, bool bTrain=false, unsigned nMaxSentSize=tagger::MAX_SENTENCE_SIZE) : m_Chart(tagger::AGENDA_SIZE) , CTaggerBase(sFeatureDBPath, bTrain, nMaxSentSize) , m_WordCache(nMaxSentSize) , m_bKnowledgeLoaded(false) { 
+   CTagger(string sFeatureDBPath, bool bTrain=false, unsigned long nMaxSentSize=tagger::MAX_SENTENCE_SIZE) : m_Chart(tagger::AGENDA_SIZE) , CTaggerBase(sFeatureDBPath, bTrain, nMaxSentSize) , m_WordCache(nMaxSentSize) , m_bKnowledgeLoaded(false) { 
       if (bTrain) m_nScoreIndex = CScore<tagger::SCORE_TYPE>::eNonAverage; else m_nScoreIndex = CScore<tagger::SCORE_TYPE>::eAverage;
    }
    virtual ~CTagger() { }
